pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "happyphone"
version = "0.1.0"
description = "E2EE encrypted voice calls and messaging - CLI version"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Happy Phone Contributors"}
]
keywords = ["e2ee", "encryption", "voip", "messaging", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "pynacl>=1.5.0",
    "python-socketio[asyncio_client]>=5.10.0",
    "aiohttp>=3.9.0",
    "aiortc>=1.6.0",
    "pyaudio>=0.2.14",
    "prompt-toolkit>=3.0.0",
    "rich>=13.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
happyphone = "happyphone.cli:main"

[project.urls]
Homepage = "https://happy.land"
Repository = "https://github.com/yourusername/happyphone-cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["happyphone*"]
